[experimental] Reduce `rpc-core` to types-only
## Summary

In this stack, we're making the decision that the `rpc-core` package should only have an opinion about _what_ the methods are (through Typescript types) and _maybe_ one day how to pre-process the inputs and post-process the server responses (but not today).

We move the responsibility for actually sending the requests to the transport.

See the top commit in this stack for how this plays out.

## Test Plan

All the tests were removed.