test: A script that you can call to start a test validator that's shared among callers ## Summary
We're about to have a situation where multiple different test runners are going to want to use a test validator. Since we can't run more than one on the same box, we're going to have to find a way for them to share one.

The idea here is that the first caller to run this script will start the validator, and the last caller to quit will stop it. We achieve this through file locks and `bash` magic.

## Test Plan

```
./scripts/start-shared-test-validator & ./scripts/start-shared-test-validator
# Ctrl-C
fg
# Ctrl-C
```