refactor(experimental): a function to patch numeric values coming from the RPC ## Summary
For the time being, we are going to need to convert numbers as they come off the RPC into `bigint`. In a follow up PR we'll allowlist certain methods/keypaths as allowed to be numbers. This will be one of the few places that we have to enumerate all RPC methods, but the damage will be minimal because _most_ values are u64.

## Test Plan

```
pnpm test:unit:browser
pnpm test:unit:node
```