refactor(experimental): add a numeric value allowlist to the response patcher
## Summary
Most of the values that come from the RPC are supposed to be `u64` (ie. must be represented as JavaScript `bigint`, eventually). Some are legitimately `usize` or `u8` though.

In this PR we create an allowlist for those method/path pairs. Any value at that path for that method will not be upcast.
## Test Plan

```
cd packages/rpc-transport
pnpm test:unit:browser
pnpm test:unit:node
```