refactor(experimental): add the `getInflationReward` RPC method ## Summary
This is a good example of a PR that adds:

* the typedef of a new RPC method
* an allowlist for numeric values
* tests

Unfortunately, parts of this method are not locally testable using currently available techniques. You can also see, in this PR, how to deal with that through the use of `it.todo()` in Jest.

## Test Plan

```
pnpm turbo test:unit:node test:unit:browser

```