refactor(experimental): coalesce requests with identical in-flight requests ## Background
Throwback to #1208. There, we implemented something I call ‘request coalescing’ at the request of @Jac0xb. He was doing a _ton_ of transaction confirmations in a long-running process, all of them using the blockheight-based confirmation strategy. This resulted in a _ton_ of `getBlockHeight` requests to their RPC infra.

We decided that because the block height depends on nothing but time, essentially, that there was no point in firing off multiple of these requests at the same time. Instead we coalesced all calls into a single request; we either started a new request or associated our request with the currently-in-fight request. This ensured that there was only one request for the same arguments in flight at any given time.

## Summary

In this PR we decide how to bring this into the rewrite of web3.js in a general way. One of this PR's central questions is: can we apply request coalescing to _every_ method?

## Discussion

### Request-in-flight based coalescing

One option is to allow an infinite number of requests to ‘get behind’ one in-flight network fetch. If no fetch exists, the first call creates it. While a fetch is in flight, the next calls add themselves as subscribers to it. The network fetch doesn't have its `AbortSignal` triggered until the last subscriber has called `abort()`.

* If a request is _not_ in flight for a call's parameters, create the request
* If a request _is_ in flight for a call's parameters, add yourself as a subscriber

*Cons*: If a request takes a long time, this might yield unexpected behavior. Imagine that a `getBlockHeight` request takes 5 _seconds_ to respond, and there are calls made at t0, t1, t2, t3, and t4. The value returned will be a different amount of stale from the perspective of each one of those calls. If the call at t4 was allowed to make its own network request, then it _might_ get fresher data.

### Time-windowed coalescing

To combat the effect of slow responses described above, we could add a rule to the coalescing algorithm:

* If a request is _not_ in flight for a call's parameters, create the request
* If a request has been in flight for a call's parameters for more than X ms, create a new request
* If a request has been in flight for a call's parameters for X or fewer ms, add yourself as a subscriber

*Cons*: While this _minimizes_ the effect of slow responses, X still represents a magic number, and each method might require a different tuning of X given its characteristics.

### JavaScript runloop-based coalescing

A variation of the time-windowed algorithm might be to make the ‘time window’ exactly the length of one JavaScript runloop. This is to say that only requests made in the _same_ runloop get coalesced together. You can think of the size of this window as ‘zero’ and the kinds of requests that it coalesces together as being the same as those typically made in a `for` loop.

*Cons*: While this will certainly reduce fetches, without the staleness problems of the last two approaches, ‘identical’ requests made very close together in time but in different runloops will not benefit from coalescing.

## Test Plan

```ts
const rpc = createSolanaRpc({ url: 'https://api.devnet.solana.com' });
for (let ii = 0; ii < 10; ii++) {
  rpc.getBlockHeight.send().then(console.log);
}

```

Observe that only one network request gets made, but all logs report the same value.