refactor(experimental): add a comparator you can use to sort base58 addresses ## Summary
We sort addresses peculiarly. Not based on the value of the underlying key. Not using the default `Array.sort()`. Rather, using `localeCompare` with a specific set of options. This PR encodes this into the new experimental library, through the use of a getter that returns you the compare functino of an `Intl.Collator`.

## Test Plan

```
pnpm test:unit:browser
pnpm test:unit:node
```