refactor(experimental): a function to compile a transaction's instructions into a message ## Summary
Over the wire, transaction instructions don't carry the addresses that they refer to, but rather the _indices_ of those addresses in a sidecar address table. This function takes in the ordered addresses and the instructions, and returns the instructions with the addresses replaced with indices.


## Test Plan

```
cd packages/transactions/
pnpm test:unit:browser
pnpm test:unit:node
```