refactor(experimental): a function to compile a transaction into an ordered list of accounts ## Summary
There is a very particular way in which accounts are laid out in Solana transactions that get sent to the network. Signers before non signers. Writable accounts before readonly accounts. Static accounts before lookup table accounts. Invoked accounts may sign transactions but may never be writable. The list goes on.

In this PR, we encode all of this logic and write copious tests for it. What results is a function that can take in a transaction (fee payer + instructions) and produce an ordered list of accounts ready to be serialized into a transactions wire format.

## Test Plan
```
cd packages/transactions/
pnpm test:unit:browser
pnpm test:unit:node
```