refactor(experimental): a seriailzer for address table lookups ## Summary
An address table lookup forms part of v0 transactions. Each one contains the address of the lookup table account itself, the indices of the accounts within that should be write-locked, and the indices of the accounts that should be read-only.

## Test Plan

```
cd packages/transactions/
pnpm test:unit:browser
pnpm test:unit:node
```