refactor(experimental): a function for generating secret keys
## Summary

This PR introduces `generateSecretKey()`. You might need to use this when you need to sign for the creation of an account, for instance.

Instead of vending the _bytes_ of a secret key, however, we vend a ‘receipt’ for it. You can return this receipt at a later time to perform some action, like deriving the public key for that secret, or signing a message.

The idea is that you can freely pass these secret keys around your application without worrying about accidentally logging the key material itself – ie. to Sentry or to the browser console.

Key material stays private to the `secrets.ts` module, stored in a `WeakMap`. When the ‘receipt’ falls out of scope, the garbage collector will reap the secret bytes as well.

## Test Plan

```
cd packages/keys/
pnpm test:unit:browser
pnpm test:unit:node
```