fix: use the native `fetch()` in Node wherever supported
# Summary

Folks are having trouble using the `node-fetch` polyfill with certain RPC providers that:

* use chunked-transfer encoding
* don't send a zero-sized chunk before closing the socket

`node-fetch` doesn't like this, and fatals.

One easy thing we can do while we figure this all out is to use the _native_ `fetch()` API included in Node 18+ by default and 17.5+ behind an experimental flag.

Fixes #1418 in Node 18+.

# Test Plan

CI run.