refactor(experimental): make sure that subscriptions clean themselves up no matter how the connection closes
# Summary

The bug here was that, if the connection closes for a non-fatal reason (eg. it was explicitly aborted) we want to unhook the subscription abort signal unconditionally. This prevents a situation where aborting a subscription after the connection has dropped runs the unsubscribe logic.

# Test Plan

```
cd packages/rpc-transport
pnpm test:unit:browser
pnpm test:unit:node
```