refactor(experimental): use `@noble/ed25519` in the polyfill (#1854)
# Summary

So OMG, the polyfill provides an implementation for the _async_ methods of `SubtleCrypto` but we were using the _sync_ methods of `@noble/curves/ed25519`, whose SHA512 implementation is implemented in _userspace_.

In this PR, we replace it with `@noble/ed25519` which drops the userspace implementation of SHA512 in favour of just using `SubtleCrypto#digest`.

The resulting bundle size of our polyfill?

| Before      | After             |
| ----------- | ----------------- |
| 15524 bytes | 6236 bytes (-59%) |

# Test Plan

```
cd packages/webcrypto-ed25519-polyfill
pnpm test:unit:browser
pnpm test:unit:node
```