Update compiled-keys.ts
I discovered a bug in the @solana/web3.js library related to the use of the logical OR assignment (||=) operator in the @solana/web3.js/lib/index.cjs.js file. 

The logical OR assignment (||=) operator is not supported in certain environments, leading to a SyntaxError. The suggested fix is to replace the ||= operator with the equivalent logical OR (||) and assignment (=) syntax, as shown below: 
Before: 
operator1 ||= operator2; 
After:
operator1 = operator1 || operator2;