Only get the cached abortable iterable factory cache error message in dev mode (#2237)
# Summary

This is the first step on the way to evicting invariant violation strings (error that should never happen, therefore would be nonsensical to build error catching around) from the production code.

Here we reorganize things so that in `__DEV__` mode the error has a message, but otherwise it gets stripped out.