Refactor decode-transaction to decode-transaction-message (#2468)
This PR updates the library helper `decodeTransaction` to a new `decodeTransactionMessage`.

Previously this took as input a serialized transaction, and returned the `Transaction` object. It used the `rpc` to look up the address lookup tables and use that to decompile the transaction.

It now takes as input a `TransactionMessage` and returns the `NewTransaction` object. It continues to use the `rpc` to look up the address lookup tables, which are passed to `decompileTransactionMessage`.

This makes it a bit less of an end-to-end helper, but I think isolates the useful functionality better. If you have a serialised transaction you can deserialize it to `NewTransaction` (message bytes + signatures) without knowing anything about lookup tables. If that's all you need, then you don't need this helper. You can also deserialize the message bytes to `CompiledTransactionMessage` without lookup tables. If you need to further convert it to the `TransactionMessage` format, then this helper will fetch the lookup tables for you and decompile it.