Add newModifyAndSignTransactions to modifying signer (#2473)
Again I've left the existing `modifyAndSignTransactions` for now because of code that uses it and hasn't yet been refactored, though this time it'll just be deleted and the new function has the correct name already. 

This is required to have the signature `NewTransaction -> NewTransaction` because of logic handling multiple modifying signers.