Refactor signers sign-transaction to use transaction messages (#2476)
The signers now all take a `NewTransaction` as input. So these functions take a transaction message (with signers) and compile it to a `NewTransaction` before sending it to the signers.

I think this PR will enable me to remove the old signer functions and clean up/rename things back, but I'll do that in a separate PR to make it easier to review.