Reset errors when changing account/chain (#2923)
# Summary

Previously, if you encountered an error (eg. some feature is not supported by some wallet / on some chain) the error message would never go away, even if you switched to a wallet that _did_ support it.

This PR resets the error boundary when the account/chain changes.

# Test Plan

1. Select Trust wallet account, which does not support `signAndSendTransaction` on devnet/testnet.
2. Switch to Phantom wallet account.

Observe that the error is cleared and the sign and send feature panel appears.