Prevent the balance component's Suspense fallback from, itself, suspending (#2934)
# Summary

Turns out, this mistake was the entire reason that I needed a `startTransition()` here where one didn't make any sense. The `fallback` here could, itself, suspend.

The only reason I stuck `<Balance />` in there was so that it would hold open the same height and not cause a layout shift. Bad move. Just draw the box the size it needs to be.

# Test Plan

Select a bunch of wallets and change the cluster. Observe the balance spinner appear and disappear without causing layout shift.