Use `RpcRequestTransformer` type parameter on request argument (#3158)
This PR fixes a previous mistake on this stack by moving the `TParam` generic type parameter from the `RpcRequest` return type to the `RpcRequest` argument.

```ts
// Before.
<TParams>(request: RpcRequest) => RpcRequest<TParams>;

// After.
<TParams>(request: RpcRequest<TParams>) => RpcRequest;
```

This is because when transforming requests, we want to assert on what we are getting. Whereas, when transforming responses, we want to assert on what we are returning. Note that this is how the parameter transformer used to work before this stack.