Combine `RpcResponseTransformer` and `RpcResponseTransformerFor` (#3162)
This PR combines the `RpcResponseTransformer` and `RpcResponseTransformerFor` types by using an optional type parameter.

Initially, `RpcResponseTransformer` was defining a generic function (as opposed to a generic type that applies its type parameter to the function). However, this was only needed in the `rpc-api` file to do:

```ts
config.responseTransformer<TypeExpectationHere>;
```

The same can be achieved using:

```ts
config.responseTransformer as RpcResponseTransformer<TypeExpectationHere>;
```

Without loosening any types as `config.responseTransformer<T>` was equivalent to a type-cast.

This makes it less tedious to define response transformers as we don't have to always cast our return type to an abstract type `T` even when we know the real return type.