Clean up SolanaRpcApi: no longer extend RpcApiMethods (#3213)
Extending `RpcApiMethods` removes the ability to access typed keys on `SolanaRpcApi` etc, because it's typed as `[method: string]: RpcApiMethod`. 

In this PR we change from:
```ts
interface GetAssetApi extends RpcApiMethods {
```

To:
```ts
type GetAssetApi = {
```

(Thanks @lorisleiva!)

This allows `createRpcApi` to maintain its typing, meaning that it still constrains any API methods we attempt to build an RPC for. But we now have typesafe `keyof SolanaRpcApi` etc. 

I've also removed the export of `RpcApiMethods`, since this should now be used only as a constraint internally and not a type externally.

I've added a typetest with a bit more detail, but this basically makes this work:
```ts
'getAccountInfo' satisfies keyof SolanaRpcApi;
// @ts-expect-error RPC API does not have this method
'someMadeUpMethod' satisfies keyof SolanaRpcApi;
```

Previously `someMadeUpMethod` would satisfy it, because `keyof SolanaRpcApi` was just string.