Make `RpcResponse` provide the resopnse payload directly (#3185)
This PR reverts some of the changes in the previous stack by making the following key change:

```ts
// Before.
type RpcResponse<TResponse = unknown> = {
    readonly json: () => Promise<TResponse>;
    readonly text: () => Promise<string>;
};

// After.
type RpcResponse<TResponse = unknown> = TResponse;
```

This is because we no longer believe that it is the responsibility of the RPC API to affect the response as text or JSON. This aspect should be encapsulated by the RPC Transport layer, meaning any custom JSON parsing/stringifying will need to be done at the RPC Transport layer.

Note that, we could simply delete the `RpcResponse` type since it returns its type parameter directly but keeping it makes the code clearer as we will use `RpcResponse` instead of `unknown` everywhere in the codebase.