Add `stringifyJsonWithBigInts` helper (#3194)
This PR adds a new `stringifyJsonWithBigInts` helper function that supports rendering `bigints` as unsafe numerical values.

This will be used — in a subsequent PR — to create a new Solana-RPC-specific HTTP transport that prevents loss of precision for large integers, in both the request and the response of the RPC cycle.