Make API plans return payloads (#3229)
This PR alters the boundaries of the RPC API layer by making the `RpcApiRequestPlan` return a `payload` of type `unknown` instead of `methodName` and `params`. This means, it is now the responsibility of the RPC API to crunch the request into a payload. The outer RPC layer now simply need to forward that `payload` to the RPC transport.