Throw when RPC API plan is missing (#3245)
This PR decouples the outer RPC layer from the concrete JSON RPC implementation by throwing an error when we cannot find an API plan for a given RPC method. Prior to this PR, we were offering a fallback API plan that transformed the RPC Request into a JSON RPC v2 payload.