Introducing channels (#3308)
# Summary

Subscriptions are different from one-shot requests in that they need two things: a ‘transport’ to fulfil requests to subscribe to some kind of notifications, but also a ‘channel’ over which to send that subscription and wait for notifications. Channels might be shared by many subscriptions, so they need to be treated as a separate type.

In this PR we create the concept of the `RpcSubscriptionsChannel` and the `RpcSubscriptionsChannelCreator`.