A channel augmenter that encodes and decodes messages as JSON (#3309)
# Summary

The new web socket channel speaks in strings/buffers. For general use in the JSON RPC Subscriptions API we need it to take in JavaScript objects, send JSON strings over the wire, and parse the strings that come back as JSON.

This is a higher-order component that wraps a `RpcSubscriptionsChannel` that speaks in strings, and makes it speak in JavaScript objects.