A channel message demultiplexer (#3311)
# Summary

In this PR, we create an efficient way of fanning notifications out to the subscribers that care about them. Instead of making _every_ subscriber responsible for transforming the response and dropping the ones it's not interested in, we make this demultiplexer transform the payload at the source, and publish it to a new channel with the subscriber id in it.