Prevent false blockheight exceedence errors when transaction confirmations are aborted (#3353)
# Summary

I noticed that my application was getting blockheight exceedence errors when there definitely was not one. Turns out the reason was because my application was aborting the confirmation, which would cause all of this code to fall through to the `BLOCKHEIGHT_EXCEEDENCE` exception.

This PR introduces a throw of the abort after every `await`. After this change, this existing text becomes consequential:

https://github.com/solana-labs/solana-web3.js/blob/e9d137651792172e16548ec9e8bbd6824c9e7189/packages/transaction-confirmation/src/__tests__/waiters-test.ts#L146-L157