Convert the websocket transport to a channel (#3314)
# Summary

This PR swaps out the old transport API for the new one based on transports and channel creators. This pushes the `AsyncIterable` creation _right_ out to the edge of the API, meaning that exactly _one_ `AsyncIterable` gets created. At all points before that the pipeline is made up of `DataPublishers`.

Among other things, this makes cleanup much more straightforward and eliminates several opportunities for memory leaks.

Closes #3090.

Or very nearly so.