Delete the cached abortable iterable factory (#3362)
# Summary

The subscriptions implementation used to be based on a chain of `AsyncIterables` which lent itself to memory leaks (see #3090). This has since been refactored in #3314, eliminating the last use of this cached abortable iterable factory code. If we ever end up bringing the concept of abortable iterables back, we'll probably do it in the way it was done in #3172.

For now, we can delete this.