Specify Node LTS as the miniumum version (#3413)
This is not a panacea, but at least it draws a line in the sand. In particular, this won't prevent people who _install_ these npm packages from having a lower Node version, but it will prevent development on the packages themselves on a Node version lower than LTS.

LTS at the time of writing was 20.18.0.

Closes #1680.