Move RpcRequest and RpcResponse types to rpc-spec-types (#3393)
One of the goals of this stack is to eventually make `RpcRequestTransformers` and `RpcResponseTransformers` usable on both RPC and RPC Subscriptions architectures. This was the original intention of the `rpc-transformers` package which already hosts request and response transformers for both but with inconsistent types.

As such, this PR aims to get closer to this goal by moving the `RpcRequest`, `RpcResponse`, `RpcRequestTransformer` and `RpcResponseTransformer` types from the `rpc-spec` package to the `rpc-spec-types` package.

As a reminder the `rpc-spec` and `rpc-subscriptions-spec` packages define the interfaces of the RPC and RPC Subscriptions architectures respectively, whereas, the `rpc-spec-types` package is a utility package that is imported by both of the former packages.

Note that, the `rpc-spec-types` package is now being exported from the main library because re-exporting that package in both `rpc` and `rpc-subscriptions` would be problematic for the main library that re-exports both.

Addresses #2983.