Use RpcRequest in createRpcMessage helper (#3394)
This PR changes the arguments of the `createRpcMessage` helper function so that it accepts an `RpcRequest` directly instead of accepting all the components inside a request (i.e. `methodName` and `params`).

We discussed this a little while ago with @mcintyre94 as, without this change, you sometimes need to express a request as an object and sometimes as its deconstructed version. This PR stack aims to embrace the `RpcRequest` object wherever applicable.