Use RpcRequest in PlanExecutor configs (#3405)
This PR replaces the `notificationName` and `params` arguments of the `RpcSubscriptionsPlanExecutor` function by a single `request` argument of type `RpcRequest`. This request argument is meant to represent the user intent to request a subscription from the RPC Subscriptions packages.

Note: A changeset for this PR is included in https://github.com/solana-labs/solana-web3.js/pull/3407.