Replace subscriptionConfigurationHash with RpcRequest in RpcSubscriptionsPlan (#3406)
This PR replaces the `subscriptionConfigurationHash` attribute from the `RpcSubscriptionsPlan` type by a `request` attribute of type `RpcRequest` representing the user intent to request to an RPC Subscription.

The main issue with having a `subscriptionConfigurationHash` attribute defined in the interfaces of RPC Subscriptions is that it is mainly useful to create a coalescer transport decorator and, therefore, is application-specific. By providing the full `RpcRequest` instead, transports can decide to use this information however they see fit. For instance, a coalescer transport may hash all of the request data whereas a sharding transport may only use the method name of the request. Notice how the `@solana/fast-stable-stringify` package has now moved from the API package to the package that defines the coalescer transport.

Additionally, this PR replaces the `getSubscriptionConfigurationHash` function of the `RpcSubscriptionsApiConfig` type by an `RpcRequestTransformer`. That way, the `createRpcSubscriptionsApi` function can prepare the raw request (from the proxy), transform it using the provided request transformer and pass the transformed request into the plan executor.