Remove `payload` and `responseTransformer` from `RpcPlan` (#3431)
This PR replaces the usage of `payload` and `responseTransformer` in favour of the new `execute` function.

Note that, this PR removes some `responseTransformer`-specific tests from the RPC tests since the outer RPC layer no longer cares about transformers. This is fine since we’re already got `responseTransformer` tests for the `createJsonRpcApi` function.

Also note that a follow-up PR will add a changeset for these changes.