Override `feePayer` signer with matching address (#3482)
This PR changes the `setTransactionMessageFeePayerSigner` helper function such that, it will always override the fee payer even when the address of the existing `TransactionSigner` matches the one provided. This is because users may want to provide a different wallet from the same address.

Addresses https://github.com/solana-labs/solana-web3.js/issues/2897.