The online/offline checker now presumes that `globalThis` is the event target (#3658)
# Summary

In the isolated Content Scripts environment, `globalThis.window` is apparently not a thing. In this PR we eliminate that nesting and presume that the event target for listening to online/offline events is `globalThis` itself.

# Test Plan

Tested that `globalThis.addEventListener('offline', () => {})` fires the callback in Brave when the network connection goes down. Presumed that the same is true for all other browsers.

Fixes #3652.